rename NewKVStore
